Move server cert generation out of header
